@@ -2,46 +2,51 @@ package com.acme.edu.iteration02;
 
 import com.acme.edu.Logger;
 import com.acme.edu.SysoutCaptureAndAssertionAbility;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 
+import static java.lang.System.lineSeparator;
+@Disabled
 public class LoggerTest implements SysoutCaptureAndAssertionAbility {
+    public static final String sep = lineSeparator();
     //region given
-    @Before
+    @BeforeEach
     public void setUpSystemOut() throws IOException {
         resetOut();
         captureSysout();
     }
 
-    @After
+    @AfterEach
     public void tearDown() {
         resetOut();
     }
     //endregion
 
 
-    /*
-    TODO: implement Logger solution to match specification as tests
+//    TODO: implement Logger solution to match specification as tests
 
     @Test
     public void shouldLogSequentIntegersAsSum() throws IOException {
         //region when
+        Logger.setSumming(true);
         Logger.log("str 1");
         Logger.log(1);
         Logger.log(2);
         Logger.log("str 2");
         Logger.log(0);
+        Logger.log();
         //endregion
 
         //region then
         assertSysoutEquals(
-            "str 1\n" +
-            "3\n" +
-            "str 2\n" +
-            "0\n"
+            "str 1" + sep +
+            "3" + sep +
+            "str 2" + sep +
+            "0" + sep
         );
         //endregion
     }
@@ -49,20 +54,22 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
     @Test
     public void shouldLogCorrectlyIntegerOverflowWhenSequentIntegers() {
         //region when
+        Logger.setSumming(true);
         Logger.log("str 1");
         Logger.log(10);
         Logger.log(Integer.MAX_VALUE);
         Logger.log("str 2");
         Logger.log(0);
+        Logger.log();
         //endregion
 
         //region then
         assertSysoutEquals(
-            "str 1\n" +
-            "10\n" +
-            Integer.MAX_VALUE + "\n" +
-            "str 2\n" +
-            "0\n"
+            "str 1" + sep +
+            "10" + sep +
+            Integer.MAX_VALUE + sep +
+            "str 2" + sep +
+            "0" + sep
         );
         //endregion
     }
@@ -70,20 +77,22 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
     @Test
     public void shouldLogCorrectlyByteOverflowWhenSequentBytes() {
         //region when
+        Logger.setSumming(true);
         Logger.log("str 1");
         Logger.log((byte)10);
         Logger.log((byte)Byte.MAX_VALUE);
         Logger.log("str 2");
         Logger.log(0);
+        Logger.log();
         //endregion
 
         //region then
         assertSysoutEquals(
-            "str 1\n" +
-            "10\n" +
-            Byte.MAX_VALUE + "\n" +
-            "str 2\n" +
-            "0\n"
+            "str 1" + sep +
+            "10" + sep +
+            Byte.MAX_VALUE + sep +
+            "str 2" + sep +
+            "0" + sep
         );
         //endregion
     }
@@ -91,6 +100,7 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
     @Test
     public void shouldLogSameSubsequentStringsWithoutRepeat() throws IOException {
         //region when
+        Logger.setSumming(true);
         Logger.log("str 1");
         Logger.log("str 2");
         Logger.log("str 2");
@@ -99,18 +109,20 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
         Logger.log("str 3");
         Logger.log("str 3");
         Logger.log("str 3");
+        Logger.log();
         //endregion
 
         //region then
         assertSysoutEquals(
-            "str 1\n" +
-            "str 2 (x2)\n" +
-            "0\n" +
-            "str 2\n" +
-            "str 3 (x3)\n"
+            "str 1" + sep +
+            "str 2 (x2)" + sep +
+            "0" + sep +
+            "str 2" + sep +
+            "str 3 (x3)" + sep
         );
         //endregion
     }
+    /*
 
     */
 }
